requirements.txt
logging==0.4.9.6
playwright==1.50.0
python-dotenv==1.0.1